Bugfix: Print CJK characters
The width of CJK characters and ASCII characters are different,
even if using the fixed font of the same size.
Moreover, one character is not necessarily
one WCHAR (UNICODE version) or char (ANSI version).
Therefore, if it printed one WCHAR or char at a time,
some characters (eg. CJK) was broken.
This patch can print 1 page at a time, and solves the problem.